refactored testing folder for datasets, minor changes
